ref(dam): Make getWidgetInterval reusable
